Upgrade to Minecraft 26.2 (Fabric + NeoForge)
Bump minecraft_version, fabric/neoforge dependency versions, Loom 1.17.11,
and Gradle 9.5.1. Fix breaking API changes: MultiBufferSource removal
(migrated to SubmitNodeCollector), Minecraft#screen/setScreen moved to Gui,
Sheets sheet method renames, BlockPos#getCenter() removal, MagmaCube/Slime
package move, and Loom includeInternal configuration removal. Both fabric
and neoforge subprojects build successfully; not yet launch-tested.
