+++ build.gradle
@@ -6,6 +6,6 @@ buildscript {
 
 plugins {
     id 'org.jetbrains.gradle.plugin.idea-ext' version '1.1.7'
-    id 'net.fabricmc.fabric-loom' version '1.16-SNAPSHOT' apply false
+    id 'net.fabricmc.fabric-loom' version '1.17.11' apply false
     id 'net.neoforged.moddev' version '2.0.141' apply false
 }
