+++ common/src/main/java/com/mrcrayfish/framework/api/client/model/renderer/StandaloneModelRenderer.java
@@ -5,7 +5,6 @@ import com.mojang.blaze3d.vertex.QuadInstance;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mrcrayfish.framework.api.client.model.FrameworkBakedModel;
 import com.mrcrayfish.framework.platform.ClientServices;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.SubmitNodeCollector;
 import net.minecraft.client.renderer.block.dispatch.BlockStateModelPart;
@@ -38,69 +37,50 @@ public class StandaloneModelRenderer
         });
     }
 
+    // changed: MultiBufferSource was removed entirely in MC 26.2 (replaced by the deferred
+    // SubmitNodeCollector pipeline) — draw(...) overloads below are now submitDraw(...) overloads
+    // taking a SubmitNodeCollector, mirroring the submitDraw(BlockStateModelPart) method above.
+
     /**
-     * Draws a FrameworkStandaloneModel into the buffer, usually in a GUI
+     * Submits a FrameworkStandaloneModel for deferred rendering, usually in a GUI
      *
+     * @param collector the SubmitNodeCollector to submit geometry to
      * @param model the model to draw
      * @param stack the current PoseStack
-     * @param source a MultiBufferSource instance
      * @param red the amount of red from 0 to 1. Only applicable if model quads are tinted
      * @param green the amount of green from 0 to 1. Only applicable if model quads are tinted
      * @param blue the amount of blue from 0 to 1. Only applicable if model quads are tinted
      * @param light the lighting for the model
      * @param overlay the overlay texture for the model
      */
-    public static void draw(FrameworkBakedModel model, PoseStack stack, MultiBufferSource source, float red, float green, float blue, int light, int overlay)
+    public static void submitDraw(SubmitNodeCollector collector, FrameworkBakedModel model, PoseStack stack, float red, float green, float blue, int light, int overlay)
     {
-        draw(model.quads(), model.layer(), stack, source, red, green, blue, light, overlay);
+        submitDraw(collector, model.quads(), model.layer(), stack, red, green, blue, light, overlay);
     }
 
     /**
-     * Draws a BlockModelPart into the buffer, usually in a GUI
+     * Submits a QuadCollection for deferred rendering, usually in a GUI
      *
-     * @param model the model to draw
-     * @param stack the current PoseStack
-     * @param source a MultiBufferSource instance
-     * @param red the amount of red from 0 to 1. Only applicable if model quads are tinted
-     * @param green the amount of green from 0 to 1. Only applicable if model quads are tinted
-     * @param blue the amount of blue from 0 to 1. Only applicable if model quads are tinted
-     * @param light the lighting for the model
-     * @param overlay the overlay texture for the model
-     */
-    public static void draw(BlockStateModelPart model, PoseStack stack, MultiBufferSource source, float red, float green, float blue, int light, int overlay)
-    {
-        VertexConsumer consumer = source.getBuffer(getSheet(ChunkSectionLayer.SOLID));
-        for(Direction direction : DIRECTIONS)
-        {
-            putQuads(stack.last(), consumer, red, green, blue, model.getQuads(direction), light, overlay);
-        }
-        putQuads(stack.last(), consumer, red, green, blue, model.getQuads(null), light, overlay);
-    }
-
-    /**
-     * Draws a QuadCollection into the buffer, usually in a GUI
-     *
+     * @param collector the SubmitNodeCollector to submit geometry to
      * @param collection the model to draw
      * @param layer the chunk section layer of the model
      * @param stack the current PoseStack
-     * @param source a MultiBufferSource instance
      * @param red the amount of red from 0 to 1. Only applicable if model quads are tinted
      * @param green the amount of green from 0 to 1. Only applicable if model quads are tinted
      * @param blue the amount of blue from 0 to 1. Only applicable if model quads are tinted
      * @param light the lighting for the model
      * @param overlay the overlay texture for the model
      */
-    public static void draw(QuadCollection collection, ChunkSectionLayer layer, PoseStack stack, MultiBufferSource source, float red, float green, float blue, int light, int overlay)
+    public static void submitDraw(SubmitNodeCollector collector, QuadCollection collection, ChunkSectionLayer layer, PoseStack stack, float red, float green, float blue, int light, int overlay)
     {
-        red = Mth.clamp(red, 0, 1);
-        green = Mth.clamp(green, 0, 1);
-        blue = Mth.clamp(blue, 0, 1);
-        VertexConsumer consumer = source.getBuffer(getSheet(layer));
-        for(Direction direction : DIRECTIONS)
-        {
-            putQuads(stack.last(), consumer, red, green, blue, collection.getQuads(direction), light, overlay);
-        }
-        putQuads(stack.last(), consumer, red, green, blue, collection.getQuads(null), light, overlay);
+        float clampedRed = Mth.clamp(red, 0, 1);
+        float clampedGreen = Mth.clamp(green, 0, 1);
+        float clampedBlue = Mth.clamp(blue, 0, 1);
+        collector.submitCustomGeometry(stack, getSheet(layer), (pose1, consumer) -> {
+            for(Direction direction : DIRECTIONS)
+                putQuads(pose1, consumer, clampedRed, clampedGreen, clampedBlue, collection.getQuads(direction), light, overlay);
+            putQuads(pose1, consumer, clampedRed, clampedGreen, clampedBlue, collection.getQuads(null), light, overlay);
+        });
     }
 
     private static void putQuads(PoseStack.Pose pose, VertexConsumer consumer, float red, float green, float blue, List<BakedQuad> quads, int light, int overlay)
@@ -122,6 +102,7 @@ public class StandaloneModelRenderer
 
     private static RenderType getSheet(ChunkSectionLayer layer)
     {
-        return layer == ChunkSectionLayer.TRANSLUCENT ? Sheets.translucentBlockSheet() : Sheets.cutoutBlockSheet();
+        // changed: Sheets.translucentBlockSheet()/cutoutBlockSheet() renamed with "Item" in MC 26.2
+        return layer == ChunkSectionLayer.TRANSLUCENT ? Sheets.translucentBlockItemSheet() : Sheets.cutoutBlockItemSheet();
     }
 }
