+++ gradle.properties
@@ -5,21 +5,21 @@ group=com.mrcrayfish
 java_version=25
 
 # Common
-minecraft_version=26.1.2
-neo_form_version=26.1.2-1
+minecraft_version=26.2
+neo_form_version=26.2-1
 
 # Fabric
-fabric_version=0.145.4+26.1.2
-fabric_loader_version=0.18.6
-fabric_minecraft_version_range=~26.1-
+fabric_version=0.152.1+26.2
+fabric_loader_version=0.19.3
+fabric_minecraft_version_range=~26.2-
 
 # NeoForge
-neoforge_version=26.1.2.4-beta
-neoforge_version_range=[26.1.2,)
+neoforge_version=26.2.0.6-beta
+neoforge_version_range=[26.2,)
 neoforge_loader_version_range=[2,)
 
 # Mod options
-mod_version=0.13.23
+mod_version=0.14.0
 mod_id=framework
 mod_name=Framework
 mod_author=MrCrayfish
